Move bins code to separate file
